Fix TypeError with nunique
